Fix issues with running Ruby 2.0.0 and using Heroku Toolbelt
The issue was that bundle sets ENV['RUBYOPT'] after it has been
loaded, which causes any subprocesses that get spawned
to be confused. If you spawn the processes using the incantation
"Bundler.with_clean_env", the RUBYOPT will be temporarily restored.

See for more info:
http://pivotallabs.com/service-oriented-foreman/

We wrote a cucumber test (shell_execution.feature) that exposes
the problem, but it requires that you have the Heroku Toolbelt
installed (not the gem) as well as RVM and Ruby 2.0.